refactor(auth-service): Split auth, geography, lookup, project, sakhi routes out of controllers
Completes the auth-service portion of the routes/controller split: all
remaining controllers (auth, geography, lookup, project, sakhi) move
route/middleware declarations and response Zod schemas into their own
*.routes.ts files. Controllers keep only request-handling logic via
createXController(service) factories. module.ts files now create the
shared DocumentedRouter and register routes into it.

geography.routes.ts preserves the original route-registration order
(roots before :id) exactly, since it is load-bearing for correct
Express path matching.

No behavior change — verified full-service OpenAPI doc (22 routes
across 6 controllers) byte-identical before/after, lint clean, all 239
existing tests pass, and live smoke checks confirming route ordering
and middleware order (validateBody before auth on /auth/login,
auth before role checks elsewhere) are preserved.
