+++ apps/auth-service/src/geography/geography.module.ts
@@ -1,15 +1,18 @@
-import type { DocumentedRouter, TokenSigner } from '@armman/service-commons';
+import type { TokenSigner } from '@armman/service-commons';
+import { createDocumentedRouter, type DocumentedRouter } from '../app.module';
 import type { PrismaService } from '../prisma/prisma.service';
 import { GeographyRepository } from './geography.repository';
 import { GeographyService } from './geography.service';
-import { createGeographyRouter } from './geography.controller';
+import { registerGeographyRoutes } from './geography.routes';
 
-/** Composition root for the geography feature: wires repository → service → router. */
+/** Composition root for the geography feature: wires repository → service → routes. */
 export function createGeographyModule(
   prisma: PrismaService,
   signer: TokenSigner,
 ): DocumentedRouter {
   const repository = new GeographyRepository(prisma);
   const service = new GeographyService(repository);
-  return createGeographyRouter(service, signer);
+  const doc = createDocumentedRouter();
+  registerGeographyRoutes(doc, service, signer);
+  return doc;
 }
